+++ src/drompa/dd_draw.cpp
@@ -343,21 +343,17 @@ void setcolor(const Cairo::RefPtr<Cairo::Context> &cr, bed &x)
 
 void setcolor(const Cairo::RefPtr<Cairo::Context> &cr, bed12 &x)
 {
+//  printf("%d, %d, %d\n", x.rgb_r, x.rgb_g, x.rgb_b);
   cr->set_source_rgba(x.rgb_r/(double)255, x.rgb_g/(double)255, x.rgb_b/(double)255, 0.6);
 }
 
-//void PDFPage::drawBedAnnotation(const vbed<auto> &vbed)
 template <class T>
-void PDFPage::drawBedAnnotation(const vbed<T> &vbed)
+void PDFPage::drawBedAnnotation(const DROMPA::Global &p, const vbed<T> &vbed)
 {
   DEBUGprint_FUNCStart();
   int32_t boxheight(BOXHEIGHT_BEDANNOTATION);
   std::string chr(rmchr(chrname));
   double ycenter(par.yaxis_now + boxheight/2);
-
-  // label
-  cr->set_source_rgba(CLR_BLACK, 1);
-  showtext_cr(cr, 90, ycenter+4, vbed.getlabel(), 12);
 
   cr->set_line_width(0.5);
   rel_xline(cr, OFFSET_X, ycenter, par.getXaxisLen());
@@ -368,15 +364,25 @@ void PDFPage::drawBedAnnotation(const vbed<T> &vbed)
   for (auto &x: vbed.getvBed()) {
     if (x.chr != chr) continue;
 
-    setcolor(cr, x);
     if (par.xstart <= x.end && x.start <= par.xend) {
+      setcolor(cr, x);
       double x1 = BP2PIXEL(x.start - par.xstart);
       double len = (x.end - x.start) * par.dot_per_bp;
       rel_xline(cr, x1, ycenter, len);
       cr->stroke();
+      if (p.drawparam.isshowbedname() && x.name != "") { // bed12
+        cr->set_source_rgba(CLR_BLACK, 1);
+        showtext_cr(cr, x1+len/2, ycenter-1, x.name, 8);
+      }
     }
   }
   cr->stroke();
+
+  // label
+  cr->set_source_rgba(CLR_BLACK, 1);
+  showtext_cr(cr, 90, ycenter+4, vbed.getlabel(), 12);
+
+
   par.yaxis_now += boxheight;
 
   DEBUGprint_FUNCend();
@@ -495,13 +501,13 @@ void PDFPage::DrawIdeogram(const DROMPA::Global &p)
 }
 
 template <class T>
-void PDFPage::Draw_vbedlist(const std::vector<vbed<T>> &vlist)
+void PDFPage::Draw_vbedlist(const DROMPA::Global &p, const std::vector<vbed<T>> &vlist)
 {
   if(!vlist.size()) return;
 
   par.yaxis_now += MERGIN_BETWEEN_READ_BED;
   for (auto &x: vlist) {
-    drawBedAnnotation(x);
+    drawBedAnnotation(p, x);
     par.yaxis_now += 2;
   }
 }
@@ -519,9 +525,10 @@ void PDFPage::StrokeEachLayer(const DROMPA::Global &p)
   // Read
   StrokeReadLines(p);
 
-  // Bed files
-  Draw_vbedlist(p.anno.vbed12list);
-  Draw_vbedlist(p.anno.vbedlist);
+  // BED12 files
+  Draw_vbedlist(p, p.anno.vbed12list);
+  // BED files
+  Draw_vbedlist(p, p.anno.vbedlist);
 
   // ChIA-Drop
   if (p.anno.existChIADrop()) {